+++ senac-cn-ado01/lib.h
@@ -1,6 +1,6 @@
 int sum(int a, int b);
 int check_representation(char *input, int base);
-int get_integer_value(char *number);
-int baseN_to_10(char *number, int base);
+long long int get_integer_value(char *number);
+long long int baseN_to_10(char *number, long long int base);
 char *convert_base(char *representation, int source_base, int destination_base);
-char *base10_to_N(int number, char *buffer,int base_n);
+char *base10_to_N(long long int number, char *buffer,int base_n);
